PB/requirements.txt
django==4.01
Pillow==9.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
requests
gdal==3.5.3
geos
postgis==1.0.4